Dovetail.Evaluate

High-level API

buildCoreFn :: Env ctx -> Module Ann -> Eval ctx (Env ctx)

Evaluate each of the bindings in a compiled PureScript module, and store the evaluated values in the environment, without evaluating any main expression.

builtIn :: ToValue ctx a => ModuleName -> Text -> a -> Env ctx

Create an environment from a Haskell value.

It is recommended that a type annotation is given for the type of the value being provided.

For example:

builtIn (ModuleName Main) "greeting" ("Hello, World!" :: Text)
builtIn (ModuleName Main) "somePrimes" ([2, 3, 5, 7, 11] :: Vector Integer)

Functions can be provided as built-ins, but the EvalT monad needs to be used to wrap any outputs (or values in positive position):

builtIn (ModuleName Main) "strip" ((pure . Text.strip) :: Text -> Eval ctx Text)
builtIn (ModuleName Main) "map" (traverse :: (Value -> Eval ctx Value) -> Vector Value -> Eval ctx (Vector Value))

Polymorphic functions can also be provided as built-ins, but values with polymoprhic types will need to be passed across the FFI boundary with monomorphic types. The type Value can always be used to represent values of unknown or polymorphic type, as in the map example above.

Evaluation

Eval/apply

eval :: forall ctx. Env ctx -> Expr Ann -> Eval ctx (Value ctx)

Evaluate a PureScript CoreFn expression in the given environment.

Note: it should not be necessary to call this function directly in most circumstances. It is provided as a helper function, for some more advanced use cases, such as setting up a custom environment.

apply :: Value ctx -> Value ctx -> Eval ctx (Value ctx)

Apply a value which represents an unevaluated closure to an argument.

Conversion to and from Haskell types

class ToValue ctx a where

Values which can be communicated across the FFI boundary from Haskell to PureScript.

Instances should identify and document any valid representations as a subset of the semantic domain Value. Such a subset can be identified by an injective function toValue, and a partial inverse, fromValue, defined on the image of toValue.

Laws:

fromValue . toValue = pure

Minimal complete definition

Nothing

Methods

toValue :: a -> Value ctx

The default implementation uses generic deriving to identify a Haskell record type with a single data constructor with a PureScript record with the same field names.

default toValue :: (Generic a, ToObject ctx (Rep a)) => a -> Value ctx

fromValue :: Value ctx -> Eval ctx a

default fromValue :: (Generic a, ToObject ctx (Rep a)) => Value ctx -> Eval ctx a

Higher-order functions

class ToValueRHS ctx a where

ToValue should support functions with types such as

a -> Eval ctx b
a -> b -> Eval ctx c
a -> b -> c -> Eval ctx d
(a -> Eval ctx b) -> Eval ctx c
(a -> b -> Eval ctx c) -> Eval ctx d

Note that every type in a return position is wrapped in the EvalT monad transformer. This is because evaluation in general may result in errors. However, a naive translation would result in too many applications of EvalT.

Specifically, we do not want to require types such as these, in which EvalT appears on the right hand side of every function arrow:

a -> Eval ctx b (b -> Eval ctx c)
a -> Eval ctx b (b -> Eval ctx (c -> Eval ctx d))

For this reason, the ToValue instance for functions delegates to this type class for the type on the right hand side of the function. It skips the application of Eval for nested function types.

Methods

toValueRHS :: a -> Eval ctx (Value ctx)

fromValueRHS :: Eval ctx (Value ctx) -> a

Foreign data types

newtype ForeignType a

This type can be used to make custom Haskell types accessible to PureScript code via the FFI's foreign import data feature.

Constructors

ForeignType

Fields

getForeignType :: a

Records

data ObjectOptions

Options for customizing generic deriving of record instances

Constructors

ObjectOptions

Fields

toPureScriptField :: Text -> Text

Map a Haskell field name to a PureScript field name on the corresponding record type.

defaultObjectOptions :: ObjectOptions

Maps Haskell field names to PureScript field names, unmodified.

genericToValue :: (Generic a, ToObject ctx (Rep a)) => ObjectOptions -> a -> Value ctx

Derived toValue function for Haskell record types which should map to corresponding PureScript record types.

genericFromValue :: (Generic a, ToObject ctx (Rep a)) => ObjectOptions -> Value ctx -> Eval ctx a

Derived fromValue function for Haskell record types which should map to corresponding PureScript record types.

class ToObject ctx f where

This class is used in the default instance for ToValue, via generic deriving, in order to identify a Haskell record type (with a single data constructor and named fields) with values in the semantic domain corresponding to a PureScript record type with the same field names.

Methods

toObject :: ObjectOptions -> f x -> HashMap Text (Value ctx)

fromObject :: ObjectOptions -> HashMap Text (Value ctx) -> Eval ctx (f x)

module Dovetail.Types

Utilities

evalPSString :: PSString -> Eval ctx Text