Dovetail.REPL

Documentation

defaultMain

Arguments

:: forall m. (MonadFix m, MonadIO m, MonadMask m)

=> Maybe ModuleName
The default module, whose members will be available unqualified in scope.

-> [ExternsFile]
Any externs files to load

-> [Ident]
Any additional identifiers which are available in the environment, but not given types in the externs file. These will be made available without type information, for debugging purposes.

-> Env m
The evaluation environment

-> m ()

Starts a minimal interactive debugger (REPL) session.

It is more likely that you will want to use the repl function to start a REPL session from within an InterpretT block.