Dovetail.Types

Evaluation

Value types

data Value ctx

The representation of values used by the interpreter - essentially, the semantic domain for a simple untyped lambda calculus with records and ADTs.

Any additional side effects which might occur in FFI calls to Haskell code are tracked by a monad in the type argument.

Constructors

Object (HashMap Text (Value ctx))
Records are represented as hashmaps from their field names to values

Array (Vector (Value ctx))

String Text

Char Char

Number Double

Int Integer

Bool Bool

Closure (Value ctx -> Eval ctx (Value ctx))
Closures, represented in higher-order abstract syntax style.

Constructor (ProperName 'ConstructorName) [Value ctx]
Fully-applied data constructors

Foreign Dynamic
Foreign data types

Evaluation monad

data Env ctx

An environment, i.e. a mapping from names to evaluated values.

An environment for a single built-in function can be constructed using the builtIn function, and environments can be combined easily using the Monoid instance for Map.

lookupEnv :: Qualified Ident -> Env ctx -> Maybe (Value ctx)

envNames :: Env ctx -> Set (Qualified Ident)

envToMap :: Env ctx -> Map (Qualified Ident) (Value ctx)

envFromMap :: Map (Qualified Ident) (Value ctx) -> Env ctx

bindEnv :: [(Qualified Ident, Value ctx)] -> Env ctx -> Env ctx

newtype Eval ctx a

The monad used by the interpreter, which supports error reporting for errors which can occur during evaluation.

The transformed monad is used to track any benign side effects that might be exposed via the foreign function interface to PureScript code.

Constructors

Eval

Fields

unEval :: ReaderT (EvaluationContext ctx) (ExceptT (EvaluationError ctx) IO) a

runEval :: ctx -> Eval ctx a -> IO (Either (EvaluationError ctx) a)

Evaluation errors

data EvaluationError ctx

An evaluation error containing the evaluation context at the point the error was raised.

Constructors

EvaluationError

Fields

errorType :: EvaluationErrorType ctx

The type of error which was raised

errorContext :: EvaluationContext ctx

The evaluation context at the point the error was raised.

data EvaluationErrorType ctx

Errors which can occur during evaluation of PureScript code.

PureScript is a typed language, and tries to prevent runtime errors. However, in the context of this interpreter, we can receive data from outside PureScript code, so it is possible that runtime errors can occur if we are not careful. This is similar to how PureScript code can fail at runtime due to errors in the FFI.

Constructors

UnknownIdent (Qualified Ident)
A name was not found in the environment

TypeMismatch Text (Value ctx)
The runtime representation of a value did not match the expected representation

FieldNotFound Text (Value ctx)
A record field did not exist in an Object value.

InexhaustivePatternMatch [Value ctx]
A pattern match failed to match its argument

InvalidNumberOfArguments Int Int
A pattern match received the wrong number of arguments

UnsaturatedConstructorApplication
A pattern match occurred against a partially-applied data constructor

InvalidFieldName PSString
A PureScript string which contains lone surrogates which could not be decoded. See PSString.

OtherError Text
An error occurred in a foreign function which is not tracked by any of the other error types.

renderEvaluationError :: RenderValueOptions -> EvaluationError ctx -> String

Render an EvaluationError as a human-readable string.

renderEvaluationErrorType :: RenderValueOptions -> EvaluationErrorType ctx -> String

Evaluation contexts

data EvaluationContext ctx

An evaluation context currently consists of an evaluation stack, which is only used for debugging purposes, plus any other domain-specific context of type ctx.

The context type is parameterized by a monad m, because stack frames can contain environments, which can in turn contain Values, which may contain monadic closures. This can be useful for inspecting values or resuming execution in the event of an error.

Constructors

EvaluationContext

Fields

callStack :: [EvaluationStackFrame ctx]

additionalContext :: ctx

Stack frames

data EvaluationStackFrame ctx

A single evaluation stack frame TODO: support frames for foreign function calls

Constructors

EvaluationStackFrame

Fields

frameEnv :: Env ctx

The current environment in this stack frame

frameSource :: SourceSpan

The source span of the expression whose evaluation created this stack frame.

frameExpr :: Expr Ann

The expression whose evaluation created this stack frame.

pushStackFrame :: Env ctx -> Expr Ann -> Eval ctx a -> Eval ctx a

Create a stack frame for the evaluation of an expression, and push it onto the stack.

throwErrorWithContext :: (MonadError (EvaluationError x) m, MonadReader (EvaluationContext x) m) => EvaluationErrorType x -> m a

Throw an error which captures the current execution context.

renderEvaluationStack :: [EvaluationStackFrame ctx] -> Text

Debugging

renderValue :: RenderValueOptions -> Value ctx -> Text

Render a Value as human-readable text.

As a general rule, apart from any closures, the rendered text should evaluate to the value you started with (when maximumDepth is not set).

data RenderValueOptions

Options when rendering values as strings using renderValue.

Constructors

RenderValueOptions

Fields

colorOutput :: Bool

Should ANSI terminal color codes be emitted

maximumDepth :: Maybe Int

The maximum depth of a subexpression to render, or Nothing to render the entire Value.

defaultTerminalRenderValueOptions :: RenderValueOptions

Some sensible default rendering options for use on a terminal which supports color.