Dovetail

Documentation

module Dovetail.Types

module Dovetail.FFI

module Dovetail.FFI

module Dovetail.FFI.Builder

High-level API

data Interpret ctx a

A monad transformer for high-level tasks involving PureScript code, including separate compilation. Its job is to keep track of available modules, any foreign imports from Haskell code, and run PureScript code.

Note: do not confuse this monad transformer with EvalT, which is only responsible for powering evaluation of PureScript expressions.

The transformed monad is used to track any benign side effects that might be exposed via the foreign function interface to PureScript code, in the same sense as EvalT.

runInterpret :: ctx -> Interpret ctx a -> IO (Either (InterpretError ctx) a)

Run a computation in the Interpret monad, possibly returning an error. Note: errors can occur during module building or evaluation (i.e. module loading).

The runInterpret function is a simpler alternative in the case where benign side-effects are not needed.

For example:

runInterpret Module () do
  -- Load the prelude
  ffi prelude
  -- Build a module from source
  build "module Main where main = \"example\"" --

runInterpret (Eval Text) () do
  ffi prelude
  _ <- build "module Main where main = \"example\""
  -- Evaluate the main function
  evalMain (ModuleName "Main")

liftEval :: Eval ctx a -> Interpret ctx a

A convenience function for running EvalT computations in Interpret, reporting errors via InterpretError.

Debugging

runInterpretWithDebugger :: ctx -> Interpret ctx a -> IO ()

Like runInterpret, but starts an interactive debugging session in the event of a debugging error.

Error messages

data InterpretError ctx

The type of errors that can occur in the Interpret monad.

Constructors

ErrorDuringEvaluation (EvaluationError ctx)
Evaluation errors can occur during the initial evaluation of the module when it is loaded into the environment.

ErrorDuringBuild BuildError
Build errors can occur if we are building modules from source or corefn.

renderInterpretError :: RenderValueOptions -> InterpretError ctx -> String

Foreign function interface

ffi :: FFI ctx -> Interpret ctx ()

Make an FFI module available for use to subsequent operations.

For example, to make the prelude available:

ffi prelude

loadEnv :: Env ctx -> Interpret ctx ()

Building PureScript source

build :: Text -> Interpret ctx (Module Ann)

Build a PureScript module from source, and make its exported functions available during subsequent evaluations.

buildCoreFn :: ExternsFile -> Module Ann -> Interpret ctx (Module Ann)

Build a PureScript module from corefn, and make its exported functions available during subsequent evaluations.

The corefn module may be preprepared, for example by compiling from source text using the functions in the Dovetail.Build module.

data BuildError

Constructors

UnableToParse (NonEmpty ParserError)

UnableToCompile MultipleErrors

InternalError

renderBuildError :: BuildError -> String

Evaluating expressions

eval

Arguments

:: ToValueRHS ctx a

=> Maybe ModuleName
The name of the "default module" whose exports will be made available unqualified to the evaluated expression.

-> Text

-> Interpret ctx (a, SourceType)

Evaluate a PureScript expression from source

evalCoreFn :: ToValueRHS ctx a => Expr Ann -> Interpret ctx a

Evaluate a PureScript corefn expression and return the result. Note: The expression is not type-checked by the PureScript typechecker. See the documentation for ToValueRHS for valid result types.

evalMain :: ToValueRHS ctx a => ModuleName -> Interpret ctx a

Evaluate main in the specified module and return the result.

newtype Eval ctx a

The monad used by the interpreter, which supports error reporting for errors which can occur during evaluation.

The transformed monad is used to track any benign side effects that might be exposed via the foreign function interface to PureScript code.

Constructors

Eval

Fields

unEval :: ReaderT (EvaluationContext ctx) (ExceptT (EvaluationError ctx) IO) a

data Env ctx

An environment, i.e. a mapping from names to evaluated values.

An environment for a single built-in function can be constructed using the builtIn function, and environments can be combined easily using the Monoid instance for Map.

runEval :: ctx -> Eval ctx a -> IO (Either (EvaluationError ctx) a)

class ToValueRHS ctx a where

ToValue should support functions with types such as

a -> Eval ctx b
a -> b -> Eval ctx c
a -> b -> c -> Eval ctx d
(a -> Eval ctx b) -> Eval ctx c
(a -> b -> Eval ctx c) -> Eval ctx d

Note that every type in a return position is wrapped in the EvalT monad transformer. This is because evaluation in general may result in errors. However, a naive translation would result in too many applications of EvalT.

Specifically, we do not want to require types such as these, in which EvalT appears on the right hand side of every function arrow:

a -> Eval ctx b (b -> Eval ctx c)
a -> Eval ctx b (b -> Eval ctx (c -> Eval ctx d))

For this reason, the ToValue instance for functions delegates to this type class for the type on the right hand side of the function. It skips the application of Eval for nested function types.

Methods

toValueRHS :: a -> Eval ctx (Value ctx)

fromValueRHS :: Eval ctx (Value ctx) -> a

class ToValue ctx a where

Values which can be communicated across the FFI boundary from Haskell to PureScript.

Instances should identify and document any valid representations as a subset of the semantic domain Value. Such a subset can be identified by an injective function toValue, and a partial inverse, fromValue, defined on the image of toValue.

Laws:

fromValue . toValue = pure

Minimal complete definition

Nothing

Methods

toValue :: a -> Value ctx

The default implementation uses generic deriving to identify a Haskell record type with a single data constructor with a PureScript record with the same field names.

default toValue :: (Generic a, ToObject ctx (Rep a)) => a -> Value ctx

fromValue :: Value ctx -> Eval ctx a

default fromValue :: (Generic a, ToObject ctx (Rep a)) => Value ctx -> Eval ctx a

REPL

repl

Arguments

:: Maybe ModuleName
The default module, whose members will be available unqualified in scope

-> Interpret ctx ()

Start an interactive debugger (REPL) session.

Re-exports

type Ann = (SourceSpan, [Comment], Maybe SourceType, Maybe Meta)

data Expr a

data Module a